@@ -38,8 +38,8 @@ public:
 	virtual void GetSavesNames( SavesNames& out_saves_names )= 0;
 
 	virtual bool SaveAvailable() const = 0;
-	virtual void SaveGame( unsigned int slot_number )= 0;
-	virtual void LoadGame( unsigned int slot_number )= 0;
+	virtual void SaveGame( uint8_t slot_number )= 0;
+	virtual void LoadGame( uint8_t slot_number )= 0;
 
 	virtual void VidRestart()= 0;
 };
