@@ -9,26 +9,60 @@ namespace SettingsKeys
 const char always_run[]= "sv_always_run";
 const char crosshair[]= "cl_crosshair";
 const char reverse_mouse[]= "in_reverse_mouse";
+const char perm_mlook[]= "in_perm_mlook";
 const char weapon_reset[]= "cl_weapon_reset";
 const char old_style_perspective[]= "cl_old_style_perspective";
 const char player_color[]= "cl_color";
 const char player_name[]= "cl_name";
-
+const char save_comment[]= "save_comment";
 const char fx_volume[]= "s_volume";
 const char cd_volume[]= "cd_volume";
 const char mouse_sensetivity[]= "cl_mouse_speed";
 const char fov[]= "cl_fov";
 
+/* essential keys */
 const char key_forward[]= "in_key_forward";
 const char key_backward[]= "in_key_backward";
 const char key_step_left[]= "in_key_left";
 const char key_step_right[]= "in_key_right";
+const char key_speed_up[]= "in_key_speed_up";
+const char key_always_run[]= "in_key_always_run";
+const char key_jump[]= "in_key_jump";
+const char key_fire[]= "in_key_fire";
+const char key_weapon_1[]= "in_key_weapon_1";
+const char key_weapon_2[]= "in_key_weapon_2";
+const char key_weapon_3[]= "in_key_weapon_3";
+const char key_weapon_4[]= "in_key_weapon_4";
+const char key_weapon_5[]= "in_key_weapon_5";
+const char key_weapon_6[]= "in_key_weapon_6";
+const char key_weapon_7[]= "in_key_weapon_7";
+const char key_weapon_8[]= "in_key_weapon_8";
+const char key_weapon_9[]= "in_key_weapon_9";
+const char key_minimap[]= "in_key_minimap";
+const char key_minimap_scale_dec[]= "in_key_minimap_scale_dec";
+const char key_minimap_scale_inc[]= "in_key_minimap_scale_inc";
+const char key_small_hud_off[]= "in_key_small_hud_off";
+const char key_small_hud_on[]= "in_key_small_hud_on";
+const char key_weapon_change[]= "in_key_weapon_change";
+const char key_weapon_next[]= "in_key_weapon_next";
+const char key_weapon_prev[]= "in_key_weapon_prev";
+const char key_perm_mlook[]= "in_key_perm_mlook"; 
+const char key_temp_mlook[]= "in_key_temp_mlook"; 
+const char key_strafe_on[]= "in_key_strafe_on";
 const char key_turn_left[]= "in_key_turn_left";
 const char key_turn_right[]= "in_key_turn_right";
 const char key_look_up[]= "in_key_look_up";
+const char key_center_view[]= "in_key_center_view";
 const char key_look_down[]= "in_key_look_down";
-const char key_jump[]= "in_key_jump";
-
+const char key_load_game[]= "in_key_load_game"; 
+const char key_save_game[]= "in_key_save_game"; 
+const char key_load_menu[]= "in_key_load_menu"; 
+const char key_save_menu[]= "in_key_save_menu"; 
+const char key_options_menu[]= "in_key_options_menu"; 
+const char key_network_menu[]= "in_key_network_menu"; 
+const char key_quit_menu[]= "in_key_quit_menu"; 
+const char key_screenshot[]= "in_key_screenshot"; 
+const char key_menu_exit[]= "in_key_menu_exit";
 const char window_width []= "r_window_width" ;
 const char window_height[]= "r_window_height";
 const char fullscreen[]= "r_fullscreen";
@@ -36,7 +70,8 @@ const char fullscreen_display[]= "r_fullscreen_display";
 const char fullscreen_width []= "r_fullscreen_width" ;
 const char fullscreen_height[]= "r_fullscreen_height";
 const char fullscreen_frequency[]= "r_fullscreen_frequency";
-
+const char vsync[]= "r_gl_vsync";
+const char draw_fps[]= "cl_draw_fps";
 const char software_rendering[]= "r_software_rendering";
 const char software_scale[]= "r_software_scale";
 
@@ -48,7 +83,6 @@ const char opengl_msaa_level[]= "r_msaa_level";
 
 const char shadows[]= "r_shadows";
 const char brightness[]= "r_brightness";
-
 } // namespace SettingsKeys
 
 } // PanzerChasm
