@@ -1,209 +1,231 @@
-#include <algorithm>
-#include <cctype>
-#include <cstring>
-
-#include "common/files.hpp"
-using namespace ChasmReverse;
-
-#include "log.hpp"
-
 #include "vfs.hpp"
 
-namespace PanzerChasm
-{
+namespace PanzerChasm {
 
-#pragma pack(push, 1)
-struct FileInfoPacked
-{
-	unsigned char name_length;
-	char name[ 12 ]; // without end null
-	unsigned int size;
-	unsigned int offset;
-};
-#pragma pack(pop)
-
-// Own replacement for nonstandard strncasecmp
-static bool StringEquals( const char* const s0, const char* const s1, const unsigned int max_length )
-{
-	unsigned int i= 0;
-	while( s0[i] != '\0' && s1[i] != '\0' && i < max_length )
+	int Vfs::SupportedFormats() const
 	{
-		if( std::tolower( s0[i] ) != std::tolower( s1[i] ) )
-			return false;
-		i++;
-	}
-
-	return i == max_length || s0[i] == s1[i];
-}
-
-static std::string ToUpper( const std::string& s )
-{
-	std::string r= s;
-	for( char& c : r )
-		c= std::toupper(c);
-	return r;
-}
-
-static const char* ExtractFileName( const char* const file_path )
-{
-	const char* file_name_pos= file_path;
-
-	const char* str= file_path;
-	while( *str != '\0' )
-	{
-		if( *str == '/' || *str == '\\' )
-			file_name_pos= str + 1u;
-
-		str++;
-	}
-
-	return file_name_pos;
-}
-
-static std::string PrepareAddonPath( const char* const addon_path )
-{
-	if( addon_path == nullptr )
-		return "";
-
-	std::string result= addon_path;
-	if( !result.empty() )
-	{
-		if( !( result.back() == '/' || result.back() == '\\' ) )
-			result.push_back( '/' );
-	}
-	return result;
-}
-
-Vfs::VurtualFileName::VurtualFileName( const char* const in_text )
-{
-	size_t i= 0u;
-	while( i < sizeof(text) && in_text[i] != '\0' )
-	{
-		text[i]= in_text[i];
-		++i;
-	}
-
-	if( i < sizeof(text) )
-		text[i]= '\0';
-}
-
-Vfs::VurtualFileName::VurtualFileName( const char* const in_text, const size_t size )
-{
-	std::memcpy( &text, in_text, std::min( size, sizeof(text) ) );
-	if( size < sizeof(text) )
-		text[size]= '\0';
-}
-
-bool Vfs::VurtualFileName::operator==( const VurtualFileName& other ) const
-{
-	return StringEquals( this->text, other.text, sizeof(text) );
-}
-
-size_t Vfs::VurtualFileNameHasher::operator()( const VurtualFileName& name ) const
-{
-	// djb2 hash
-	size_t hash= 5381u;
-	for( size_t i= 0u; i < sizeof(name.text) && name.text[i] != '\0'; ++i )
-		hash= hash * 33u + size_t(std::tolower(name.text[i]));
-
-	return hash;
-}
-
-Vfs::Vfs(
-	const char* archive_file_name,
-	const char* const addon_path )
-	: archive_file_( std::fopen( archive_file_name, "rb" ) )
-	, addon_path_( PrepareAddonPath( addon_path ) )
-{
-	if( archive_file_ == nullptr )
-	{
-		Log::FatalError( "Could not open file \"", archive_file_name, "\"" );
-		return;
-	}
-
-	char header[4];
-	FileRead( archive_file_, header, sizeof(header) );
-	if( std::strncmp( header, "CSid", sizeof(header) ) != 0 )
-	{
-		Log::FatalError( "File \"", archive_file_name, "\" is not \"Chasm: The Rift\" archive" );
-		return;
-	}
-
-	unsigned short files_in_archive_count;
-	FileRead( archive_file_, &files_in_archive_count, sizeof(files_in_archive_count) );
-
-	std::vector<FileInfoPacked> files_info_packed( files_in_archive_count );
-	FileRead( archive_file_, files_info_packed.data(), files_info_packed.size() * sizeof(FileInfoPacked ) );
-
-	virtual_files_.reserve( files_in_archive_count );
-
-	for( const FileInfoPacked& file_info_packed : files_info_packed )
-	{
-		VirtualFile file;
-		std::memcpy( &file.size  , &file_info_packed.size  , sizeof(unsigned int) );
-		std::memcpy( &file.offset, &file_info_packed.offset, sizeof(unsigned int) );
-
-		virtual_files_[ VurtualFileName( file_info_packed.name, file_info_packed.name_length ) ]= file;
-	}
-}
-
-Vfs::~Vfs()
-{
-	if( archive_file_ != nullptr )
-		std::fclose( archive_file_ );
-}
-
-Vfs::FileContent Vfs::ReadFile( const char* const file_path ) const
-{
-	FileContent result;
-	ReadFile( file_path, result );
-	return result;
-}
-
-void Vfs::ReadFile( const char* const file_path, FileContent& out_file_content ) const
-{
-	const char* const file_name= ExtractFileName( file_path );
-	if( file_name[0] == '\0' )
-	{
-		// Do not load files with empty path.
-		out_file_content.clear();
-		return;
-	}
-
-	// Try read from real file system.
-	if( !addon_path_.empty() )
-	{
-		std::string fs_file_path= addon_path_ + ToUpper(file_path); // Use ToUpper, because files in addons are in upper case.
-		std::replace(fs_file_path.begin(), fs_file_path.end(), '\\', '/' ); // Change shitty DOS/Windows path separators to universal windows/unix separators.
-
-		std::FILE* const fs_file= std::fopen( fs_file_path.c_str(), "rb" );
-
-		if( fs_file != nullptr )
+		if( PHYSFS_isInit() == 0 )
 		{
-			std::fseek( fs_file, 0, SEEK_END );
-			const unsigned int file_size= std::ftell( fs_file );
-			std::fseek( fs_file, 0, SEEK_SET );
+			if( PHYSFS_init( "" ) == 0 )
+			{
+				Log::FatalError( "PHYSFS_init() failed!\n reason %s.\n", PHYSFS_getErrorByCode(PHYSFS_getLastErrorCode() ) );
+				std::error_code ec;
+				throw std::filesystem::filesystem_error(std::string( "Could not initialize PhysFS!" ), ec);
+			}
+		}
 
-			out_file_content.resize( file_size );
-			FileRead( fs_file, out_file_content.data(), file_size );
+		const PHYSFS_ArchiveInfo **rc = PHYSFS_supportedArchiveTypes();
+		const PHYSFS_ArchiveInfo **i;
+		int supported = DIR;
 
-			std::fclose( fs_file );
-			return;
+		if( *rc == nullptr )
+			printf( "PHYSFS BIN/TAR format not supported!\n" );
+		else
+		{
+			for(i = rc; *i != nullptr; i++)
+			{
+				if( memcmp((*i)->extension, "TAR", 3) == 0 )
+					supported |= TAR;
+				if( memcmp((*i)->extension, "BIN", 3) == 0 )
+					supported |= CSM;
+			}
+		}
+		return supported;
+	}
+
+	int Vfs::Type( const std::filesystem::path& path ) const
+	{
+		static const char* CSM_MAGIC = "CSid";
+		static const char* TAR_MAGIC = "ustar  ";	/* 7 chars and a null */
+		static const ssize_t TAR_MAGIC_OFFSET = 257;
+		static const ssize_t TAR_BLOCK_SIZE = 512;
+		static const ssize_t CSM_BLOCK_SIZE = 4;
+		int res = NIL;
+		if( !path.empty() )
+		{ 
+			if( !std::filesystem::is_directory( path ) )
+			{
+				char buf[TAR_BLOCK_SIZE + 1] = { '\0' };
+				std::ifstream is( path, std::ios_base::binary );
+				is.read( buf, TAR_BLOCK_SIZE );
+				if( is.tellg() == TAR_BLOCK_SIZE )
+				{
+					if( strncmp( buf, CSM_MAGIC, CSM_BLOCK_SIZE ) == 0 )
+					{
+						res = CSM;
+					}
+					else if( strncmp(&buf[TAR_MAGIC_OFFSET], TAR_MAGIC, CSM_BLOCK_SIZE + 1 ) == 0 )
+						res = TAR;
+				}
+				is.close();
+			}
+			else
+			{
+				res = DIR;
+			}
+		}
+		return res;
+	}
+
+	Vfs::Vfs( const std::filesystem::path& archive_file, const std::filesystem::path& addon_path )
+	{
+		archive_.path = archive_file;
+		archive_.type = (enum flags)Type( archive_.path );
+		addon_.path   = addon_path;
+		addon_.type   = (enum flags)Type( addon_.path );
+		formats_ = SupportedFormats();
+
+		if( !archive_.path.empty() )
+		{
+			if( (PHYSFS_mount( archive_.path.c_str(), "/", 0 ) == 0) )
+			{
+				Log::FatalError( "Could not open archice file \"", archive_.path, "\"" );
+				PHYSFS_deinit();
+				std::error_code ec;
+				throw std::filesystem::filesystem_error( std::string( "Could not initialize PhysFS!" ), ec );
+			}
+			Log::Info( archive_.type == TAR ? "TAR" : "BIN", " ", archive_.path );
+		}
+
+		if( !addon_.path.empty() )
+		{
+			if( ( PHYSFS_mount( addon_.path.c_str(), "/", 0 ) == 0) )
+			{
+				Log::FatalError( "Could not open addon path \"", addon_.path, "\"" );
+				PHYSFS_deinit();
+				std::error_code ec;
+				throw std::filesystem::filesystem_error( std::string( "Could not initialize PhysFS!" ), ec);
+			}
+			Log::Info( addon_.type == DIR ? "DIR" : "NIL", " ", addon_.path );
 		}
 	}
 
-	const auto it= virtual_files_.find( VurtualFileName( file_name ) );
-	if( it != virtual_files_.end() )
+	Vfs::~Vfs()
 	{
-		const VirtualFile& file= it->second;
-		out_file_content.resize( file.size );
-		std::fseek( archive_file_, file.offset, SEEK_SET );
-		FileRead( archive_file_, out_file_content.data(), out_file_content.size() );
-
-		return;
+		PHYSFS_deinit();
 	}
 
-	out_file_content.clear();
-}
+	Vfs::FileContent Vfs::ReadFile( const std::filesystem::path& file_path ) const
+	{
+		FileContent result;
+		ReadFile( file_path, result );
+		return result;
+	}
 
-} // namespace PanzerChasm
+	void Vfs::ReadFile( const std::filesystem::path& file_path, FileContent& out_file_content ) const
+	{	
+		std::string file_path_str = file_path.native();
+		std::replace(file_path_str.begin(), file_path_str.end(), '\\', '/');
+		std::filesystem::path loc( ToUpper( file_path_str ) );
+
+		if( !loc.empty() )
+		{
+			std::error_code ec;
+			if( PHYSFS_isInit() == 0 )
+			{
+				throw std::filesystem::filesystem_error( std::string( "Could not initialize PhysFS!" ), ec );
+			}
+
+			if( !addon_.path.empty() && addon_.type == DIR )
+			{
+				PHYSFS_File *f = PHYSFS_openRead( loc.c_str() );
+				if( f != nullptr )
+				{
+					PHYSFS_sint64 file_size= PHYSFS_fileLength( f );
+					out_file_content.resize( file_size );
+					PHYSFS_sint64 read = PHYSFS_readBytes( f, out_file_content.data(), out_file_content.size() );
+					if( read == out_file_content.size() )
+					{
+						PHYSFS_flush( f );
+						PHYSFS_close( f );
+						return;
+					}
+					out_file_content.clear();
+				}
+				PHYSFS_close(f);
+			}
+
+			if( !archive_.path.empty() && archive_.type == CSM )
+			{
+				PHYSFS_File *f = PHYSFS_openRead( loc.filename().c_str()  );
+				if( f != nullptr )
+				{
+					PHYSFS_sint64 file_size = PHYSFS_fileLength( f );
+					out_file_content.resize( file_size );
+					PHYSFS_sint64 read = PHYSFS_readBytes( f, out_file_content.data(), out_file_content.size() );
+					if( read == out_file_content.size() )
+					{
+						PHYSFS_flush( f );
+						PHYSFS_close( f );
+						return;
+					}
+					out_file_content.clear();
+				}
+				PHYSFS_close( f );
+			}
+
+			if( !archive_.path.empty() && archive_.type == TAR )
+			{
+				loc = ToLower( loc.native() );
+				if( loc.filename().extension() == ".raw" || loc.filename().extension() == ".sfx" || loc.filename().extension() == ".pcm" )
+				{
+					if( !(loc.filename() == "aproc.sfx" || loc.filename() == "wswap.sfx" || loc.filename() == "thund3.raw" || loc.filename() == "volc.raw" || loc.filename() == "step02.raw" ) )
+						loc = "sound" / loc.filename().replace_extension( ".wav" );
+				}
+				else if( loc.filename().extension() == ".wav" )
+					loc = "sound" / loc.filename();
+				else if( loc.filename().extension() == ".cel" )
+					loc = "texture" / loc.filename(); 
+				else if( loc.parent_path() == "models" )
+					loc = "model" / loc.filename(); 
+				else if( loc.filename().extension() == ".car" )
+					loc = "monster" / loc.filename();
+				else if( loc.parent_path() == "ani" || loc.parent_path() == "ani/weapon" || loc.parent_path() == "ani/weapon" )
+					loc = "model" / loc.filename();
+				else if( loc.stem() == "floors" || loc.stem() == "resource" || loc.stem() == "process" || loc.stem() == "map" || loc.stem() == "script" )
+					loc = "map" / loc.filename();
+				else if( loc.filename().extension() == ".obj" )
+				{
+					if( loc.parent_path() != "monster" )
+						loc = "sprite_effects" / loc.filename();
+				}
+				else if( loc.filename().extension() == ".ani" )
+					loc = "model" / loc.filename();
+				if( loc.filename().extension() == ".3o" )
+					loc = "model" / loc.filename();
+				else if( loc.filename().extension() == ".gltf" )
+				{
+					if( loc.parent_path() != "monster" )
+						loc = "model" / loc.filename();
+				} 
+
+				if( !loc.filename().has_extension() )
+				{
+					std::filesystem::path oldloc = loc;
+					if( loc.filename() == "faust" )
+						loc = "monster" / loc.filename().replace_filename( "faust1.car" );
+					if( loc.filename() == "vent1" )
+						loc = "model" / loc.filename().replace_filename( "vent1.gltf" );
+					Log::Warning( "Warning: ", oldloc, " -> ", loc, " renamed!" );
+				}
+
+				char* path = (char*)loc.c_str();
+				PHYSFS_File *f = PHYSFS_openRead( ( path ) );
+				if( f != nullptr )
+				{
+					PHYSFS_sint64 file_size = PHYSFS_fileLength( f );
+					out_file_content.resize( file_size );
+					PHYSFS_sint64 read = PHYSFS_readBytes( f, out_file_content.data(), out_file_content.size() );
+					if( read == out_file_content.size() )
+					{
+						PHYSFS_flush( f );
+						PHYSFS_close( f );
+						return;
+					}
+					out_file_content.clear();
+				}
+				PHYSFS_close( f );
+			}
+		}
+	}
+};
